Support serializing index-based types as maps as well as arrays
Arrays are most compact when they are packed with data. But if there are 'holes' in the indexes (whether by missing indexes in the key attributes or by properties having default values and thus not serialized), a map that uses those same ordinal indexes as keys can be more compact.
With this new feature, the more compact of the two formats is automatically chosen (when `SerializeDefaultValues == false`) during serialization.