Implement stateful, incremental skip
This is a perf optimization for streaming scenarios where a converter needs to skip a structure but the buffer may not have the entire structure in memory. This optimization records how much of the structure has been skipped already so that the reader can advance, freeing memory, and resume skipping when we get more bytes to decode.

In particular, this avoids the previous behavior that skipping had to be restarted from the beginning of the structure each time new bytes are brought in until the whole structure was in memory together.

While this is theoretically a memory use improvement (for skipping scenarios only), this is mostly about avoiding CPU work to parse partial structures repeatedly until we get the whole thing in memory. With this change, at most one msgpack *token* (rather than whole structure) will be decoded more than once, and that limited to only once per fetch of new bytes.

Closes #154