Allow more effective trimming
Instead of instantiating all built-in converters on startup unconditionally or referencing all the supported data types themselves, use a generic method so that the JIT will remove references to data types (and their converters) that are never serialized in the application.

It may not be intuitive, but indeed a long C# method of type checks _does_ offer this trim optimization, because the JIT / AOT compiler will remove type checks that will never match given the type argument on the generic method, thereby removing the unrelated type references, leaving only those behind that are actually going to be used.

Accrues toward #190, by bringing the AOT console test app down from 3.9MB to just under 3.6MB.