Change the default policy to include required properties at their default values
This causes C# required values to be serialized, which lends to them being present for *deserialization*, which is fitting since when coding, the value must be provided, so serialized forms ought to include them as well.
This can of course be overridden back to the old default value of `Never`.

This also has the nice effect of causing the generated JSON schemas to indicate which properties are required consistent with their C# definitions, as proposed in #133.