Fix `MessagePackStreamingReader` multi-read methods in insufficient buffer scenarios
Several read methods read a header and then the data in two steps as an implementation detail.
When there is sufficient buffer to read the first time, but insufficient the second time, these methods were all failing to "rewind" the reader so that buffering more and retrying could be done from the original point.
These bugs led to a `TokenMismatch` error on subsequent attempts.

Fixed APIs are:

- `MessagePackStreamingReader.TryRead(out Extension)`
- `MessagePackStreamingReader.TryReadBinary`
- `MessagePackStreamingReader.TryRead(out string)`
- `MessagePackStreamingReader.TryReadStringSequence`
- `MessagePackStreamingReader.TryReadStringSpan`