Make reader and writers ordinary structs
Instead of `MessagePackReader` and `MessagePackWriter` each being a `ref struct`, they are now just `struct`.
This should make certain things simpler in an async world. But perhaps more important, it means that these can be used as generic type arguments, opening the door for multiple formats of readers/writers that are compiler optimized/inlined.