Add deserialization guard for objects that record property values multiple times
This is a well known attack vector, where an object may define the value for a property multiple times, in order to get one security gate to see one value and approve, knowing that a later data processor will see the other value and fall victim.
As mitigation, we simply throw any time we deserialize a value for a given property more than once for a given object.