Throw on serializing `DateTimeKind.Unspecified`
MessagePack-CSharp suffers from accidental datetime corruption for `Unspecified` by assuming it is UTC. If it was actually local time, it is corrupted.
To avoid this being an issue for the new Nerdbank.MessagePack library, just throw when that ambiguity is encountered, forcing users to specify which Kind it is.

To keep it amiable though, we add an optional configuration where the application can specify which `DateTimeKind` to assume when serialized values are left `Unspecified`.
This also helps us workaround https://github.com/eiriktsarpalis/PolyType/issues/190.