Revise union-related API names
- `DerivedTypeMappingCollection` -> `DerivedTypeUnionCollection`
- `DerivedTypeMapping` -> `DerivedTypeUnion`
- `MessagePackSerializer.DerivedTypeMappings` -> `MessagePackSerializer.DerivedTypeUnions`

This is in preparation for union support that is not a simple _mapping_ between types and type identifiers.

In the future, support for *discriminated* type unions will be something else altogether because their union cases are not all assignable to some base type. PolyType will support these and we will too, but likely this will _not_ be via some runtime-defined union.