Avoid JIT costs incurred by primitive type converter checks
On .NET Framework, we preserve avoiding assembly loads while removing the JIT time involved with the generic methods.
On .NET, we want to preserve trimmability so we keep these methods generic, at the potential cost of more JIT time.

This tuning has no public API impact so it's something we can tune over time.

One thing we might do to reduce JIT on .NET for example is add a non-generic method that simply tests whether the type to be converted is one of the supported primitives. That would ensure that the generic instantiation only occurs when it will be worthwhile.