Avoid the large object heap for async deserialization
`StreamPipeReader.ReadAtLeastAsync` is unfortunately written to allocate a single, contiguous buffer of the requested `minimumSize` regardless of whether it is of an acceptable size per the memory pool configuration.
When we want to read that much, it isn't that we need it to be in contiguous memory, and we want to avoid the LOH. So we now call our own helper method that fills the same need but without allocating absurd buffer sizes.

This is a workaround for https://github.com/dotnet/runtime/issues/120618