Change `SerializeDefaultValues` default to `Always`
Skipping properties with 'default values' is problematic because we don't know the default value when we consider initializers and constructors that change the default from `default(T)`. This can lead to counter-intuitive round-tripping failures and data loss.

Skipping default value serialization should be opt-in so that developers have a chance to read the docs explaining its requirements and assumptions.

Fixes #778